KD-3172 - (squashable) Slowness in acquisition
The translation tool completely messes up the HTML for the element
where title attributes have translatable text and TT variables, and
the titles are inside a string, like in this JSON output.

Solution: Put the translatable title attribute texts into variables
outside the HTML element, and use the variable in the title attribute.